requirements.txt
# Core dependencies
openai==0.27.8
deepgram-sdk==2.12.0
aiohttp==3.8.5
aiofiles==23.1.0
requests==2.31.0

# Data processing
pandas==2.0.3
numpy==1.24.3
pydantic==2.1.1

# Audio processing
librosa==0.10.1
soundfile==0.12.1

# Visualization and analysis
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
rich==13.5.2
click==8.1.6

# Async support
asyncio
uvloop==0.17.0

# Development tools (optional)
pytest==7.4.0
black==23.7.0
flake8==6.0.0